server/requirements.txt
amqp==5.0.6
billiard==3.6.4.0
bokeh==2.3.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudpickle==1.6.0
colorcet==2.0.6
cycler==0.10.0
Cython==0.29.23
dask==2021.4.1
datashader==0.12.1
datashape==0.5.2
distributed==2021.4.1
Flask==1.1.2
Flask-Cors==3.0.10
fsspec==2021.4.0
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
hdbscan==0.8.27
HeapDict==1.0.1
holoviews==1.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.0.2
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
msgpack==1.0.2
multipledispatch==0.6.0
numba==0.53.1
numpy==1.20.2
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.4
panel==0.11.3
param==1.10.1
partd==1.2.0
Pillow==8.2.0
prompt-toolkit==3.0.18
psutil==5.8.0
pyct==0.4.8
pynndescent==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pyviz-comms==2.0.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sortedcontainers==2.3.0
tblib==1.7.0
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
typing-extensions==3.10.0.0
umap-learn==0.5.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
xarray==0.18.0
zict==2.0.0
zope.event==4.5.0
zope.interface==5.4.0
